pyproject.toml
[tool.poetry]
name = "tofino-traffic-generator"
version = "0.1.0"
description = ""
authors = ["Pascal Bast <pascal.bast@student.htw-berlin.de>"]
readme = "README.md"

[tool.poetry.dependencies]
bfrt-helper = { git = "https://github.com/APS-Networks/bfrt-helper.git", branch = "main" }
python = "^3.8"
googleapis-common-protos = "1.63.2"
grpcio = "1.65.1"
grpcio-tools = "1.65.1"
protobuf = "5.27.2"
setuptools = "71.0.3"
paramiko = "^3.4.0"
jinja2 = "^3.1.4"
black = "^24.4.2"
grpcio-health-checking = "^1.65.1"
scapy = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
